+++ client/src/components/common/Header/SearchInput.tsx
@@ -1,7 +1,6 @@
 import React, { useRef, KeyboardEvent, useState, ChangeEvent } from 'react';
-import styled from '@emotion/styled';
-import { css } from '@emotion/react';
-import FONT from '@/styles/font';
+import * as Styled from './SearchInput.styles';
+
 import SearchIcon from '@/assets/icon/SearchIcon.svg';
 
 export interface SearchInputProps {
@@ -39,9 +38,9 @@ export default function SearchInput(props: SearchInputProps): JSX.Element {
   };
 
   return (
-    <Container active={active} onClick={onClick}>
-      <Text>검색</Text>
-      <Input
+    <Styled.Container active={active} onClick={onClick}>
+      <Styled.Text>검색</Styled.Text>
+      <Styled.Input
         data-testid="search-input"
         type="text"
         placeholder={placeholder}
@@ -49,76 +48,9 @@ export default function SearchInput(props: SearchInputProps): JSX.Element {
         onChange={handleClick}
         onKeyDown={handleKeydown}
       />
-      <Icon data-testid="search-icon" onClick={handleSubmit}>
+      <Styled.Icon data-testid="search-icon" onClick={handleSubmit}>
         <SearchIcon />
-      </Icon>
-    </Container>
+      </Styled.Icon>
+    </Styled.Container>
   );
 }
-
-interface ContainerProps {
-  active: boolean;
-}
-
-const activeCSS = css`
-  width: 550px;
-  padding: 0 12px;
-  border: 1px solid #cdcdcd;
-  background-color: #ffffff;
-  span:first-of-type {
-    visibility: hidden;
-  }
-  input {
-    visibility: visible;
-    width: 500px;
-  }
-`;
-
-const Container = styled.div<ContainerProps>`
-  display: flex;
-  width: 60px;
-  height: 30px;
-  margin: 18px;
-  border: 0 solid #cdcdcd;
-  border-radius: 8px;
-  background-color: transparent;
-  align-items: center;
-  transition: width 0.35s;
-
-  input {
-    visibility: hidden;
-    width: 0;
-    padding: 0;
-    margin: 0;
-    transition: width 0.35s;
-  }
-  &:hover,
-  &:focus,
-  &:focus-within {
-    ${activeCSS}
-  }
-  ${({ active }) => active && activeCSS}
-`;
-
-const Text = styled.span`
-  position: absolute;
-  white-space: nowrap;
-`;
-
-const Input = styled.input`
-  position: absolute;
-  flex: 1;
-  border: none;
-  outline: none;
-  ${FONT.body4};
-
-  ::placeholder {
-    color: #cdcdcd;
-  }
-`;
-
-const Icon = styled.span`
-  margin-top: 4px;
-  margin-left: auto;
-  cursor: pointer;
-`;
